test: extract breadcrumbs overflow unit tests to separate file
