+++ packages/breadcrumbs/test/breadcrumbs.test.js
@@ -1,8 +1,6 @@
 import { expect } from '@vaadin/chai-plugins';
-import { sendKeys } from '@vaadin/test-runner-commands';
-import { fixtureSync, nextRender, nextResize, oneEvent } from '@vaadin/testing-helpers';
+import { fixtureSync, nextRender } from '@vaadin/testing-helpers';
 import '../vaadin-breadcrumbs.js';
-import { getDeepActiveElement } from '@vaadin/a11y-base/src/focus-utils.js';
 
 window.Vaadin ??= {};
 window.Vaadin.featureFlags ??= {};
@@ -168,245 +166,6 @@ describe('vaadin-breadcrumbs', () => {
 
       const remaining = [...breadcrumbs.querySelectorAll('vaadin-breadcrumbs-item')];
       expect(remaining[0].getAttribute('slot')).to.equal('root');
-    });
-  });
-
-  describe('overflow', () => {
-    let items, button, overlay;
-
-    function expectOverlayItems(expectedIndexes) {
-      items.forEach((item, index) => {
-        const inOverlay = item.slot === 'overlay';
-        const shouldBeInOverlay = expectedIndexes.includes(index);
-        expect(inOverlay, `item ${index}`).to.equal(shouldBeInOverlay);
-      });
-    }
-
-    beforeEach(async () => {
-      breadcrumbs = fixtureSync(`
-        <vaadin-breadcrumbs style="max-width: 800px;">
-          <vaadin-breadcrumbs-item path="/">Home</vaadin-breadcrumbs-item>
-          <vaadin-breadcrumbs-item path="/docs">Documents</vaadin-breadcrumbs-item>
-          <vaadin-breadcrumbs-item path="/docs/projects">Projects</vaadin-breadcrumbs-item>
-          <vaadin-breadcrumbs-item path="/docs/projects/2026">2026</vaadin-breadcrumbs-item>
-          <vaadin-breadcrumbs-item path="/docs/projects/2026/q1">Quarter Reports</vaadin-breadcrumbs-item>
-          <vaadin-breadcrumbs-item>Summary report</vaadin-breadcrumbs-item>
-        </vaadin-breadcrumbs>
-      `);
-      await nextRender();
-      items = [...breadcrumbs.querySelectorAll('vaadin-breadcrumbs-item')];
-      // Pin each item to a fixed width so collapse thresholds in the
-      // tests below are not sensitive to font metric differences
-      // between local macOS and CI Linux.
-      items.forEach((item) => {
-        item.style.width = '100px';
-      });
-      await nextResize(breadcrumbs);
-      button = breadcrumbs.shadowRoot.querySelector('[part="overflow-button"]');
-      overlay = breadcrumbs.shadowRoot.querySelector('vaadin-breadcrumbs-overlay');
-    });
-
-    describe('detection', () => {
-      it('should collapse the closest-to-root item first', async () => {
-        breadcrumbs.style.maxWidth = '600px';
-        await nextResize(breadcrumbs);
-
-        expectOverlayItems([1]);
-      });
-
-      it('should keep the root in the trail until last', async () => {
-        breadcrumbs.style.maxWidth = '600px';
-        await nextResize(breadcrumbs);
-
-        expect(items[0].slot).to.equal('root');
-      });
-
-      it('should never collapse the last item', async () => {
-        breadcrumbs.style.maxWidth = '20px';
-        await nextResize(breadcrumbs);
-
-        expect(items.at(-1).slot).to.not.equal('overlay');
-      });
-
-      it('should collapse more items as the container shrinks', async () => {
-        breadcrumbs.style.maxWidth = '600px';
-        await nextResize(breadcrumbs);
-        expectOverlayItems([1]);
-
-        breadcrumbs.style.maxWidth = '280px';
-        await nextResize(breadcrumbs);
-        expectOverlayItems([1, 2, 3, 4]);
-      });
-
-      it('should collapse the root when only the last item fits', async () => {
-        breadcrumbs.style.maxWidth = '150px';
-        await nextResize(breadcrumbs);
-
-        expect(items[0].slot).to.equal('overlay');
-      });
-
-      it('should restore items when the container widens', async () => {
-        breadcrumbs.style.maxWidth = '280px';
-        await nextResize(breadcrumbs);
-
-        breadcrumbs.style.maxWidth = '800px';
-        await nextResize(breadcrumbs);
-
-        expectOverlayItems([]);
-      });
-
-      it('should clear has-overflow when items fit again', async () => {
-        breadcrumbs.style.maxWidth = '280px';
-        await nextResize(breadcrumbs);
-
-        breadcrumbs.style.maxWidth = '800px';
-        await nextResize(breadcrumbs);
-
-        expect(breadcrumbs.hasAttribute('has-overflow')).to.be.false;
-      });
-
-      it('should re-run detection when items are added', async () => {
-        for (let i = 0; i < 6; i += 1) {
-          const item = document.createElement('vaadin-breadcrumbs-item');
-          item.path = `/extra/${i}`;
-          item.textContent = `Extra long item label ${i}`;
-          breadcrumbs.insertBefore(item, breadcrumbs.lastElementChild);
-        }
-        await nextRender();
-
-        expect(breadcrumbs.hasAttribute('has-overflow')).to.be.true;
-      });
-    });
-
-    describe('overlay', () => {
-      it('should set overlay owner to the breadcrumbs element', () => {
-        expect(overlay.owner).to.equal(breadcrumbs);
-      });
-
-      it('should set overlay positionTarget to the overflow button', () => {
-        expect(overlay.positionTarget).to.equal(button);
-      });
-    });
-
-    describe('opening', () => {
-      beforeEach(async () => {
-        breadcrumbs.style.maxWidth = '200px';
-        await nextResize(breadcrumbs);
-      });
-
-      it('should toggle overlay opened on button click', async () => {
-        button.click();
-        await nextRender();
-
-        expect(overlay.opened).to.be.true;
-
-        button.click();
-        await nextRender();
-
-        expect(overlay.opened).to.be.false;
-      });
-
-      it('should update aria-expanded attribute on the overflow button', async () => {
-        button.click();
-        await nextRender();
-
-        expect(button.getAttribute('aria-expanded')).to.equal('true');
-
-        button.click();
-        await nextRender();
-
-        expect(button.getAttribute('aria-expanded')).to.equal('false');
-      });
-
-      it('should open the overlay when Enter is pressed', async () => {
-        button.focus();
-        await sendKeys({ press: 'Enter' });
-        await nextRender();
-
-        expect(overlay.opened).to.be.true;
-      });
-
-      it('should open the overlay when Space is pressed', async () => {
-        button.focus();
-        await sendKeys({ press: 'Space' });
-        await nextRender();
-
-        expect(overlay.opened).to.be.true;
-      });
-
-      it('should move focus to the first overlay item after opening via Enter', async () => {
-        button.focus();
-        await sendKeys({ press: 'Enter' });
-        await nextRender();
-
-        const firstItem = breadcrumbs.querySelector('vaadin-breadcrumbs-item[slot="overlay"]');
-        const firstLink = firstItem.shadowRoot.querySelector('[part="link"]');
-        expect(getDeepActiveElement()).to.equal(firstLink);
-      });
-
-      it('should move focus to the first overlay item after opening via Space', async () => {
-        button.focus();
-        await sendKeys({ press: 'Space' });
-        await nextRender();
-
-        const firstItem = breadcrumbs.querySelector('vaadin-breadcrumbs-item[slot="overlay"]');
-        const firstLink = firstItem.shadowRoot.querySelector('[part="link"]');
-        expect(getDeepActiveElement()).to.equal(firstLink);
-      });
-    });
-
-    describe('closing', () => {
-      beforeEach(async () => {
-        breadcrumbs.style.maxWidth = '600px';
-        await nextResize(breadcrumbs);
-
-        button.click();
-        await oneEvent(overlay, 'vaadin-overlay-open');
-      });
-
-      it('should close the overlay on Escape key press', async () => {
-        await sendKeys({ press: 'Escape' });
-        await nextRender();
-
-        expect(overlay.opened).to.be.false;
-      });
-
-      it('should close the overlay on Tab key press', async () => {
-        await sendKeys({ press: 'Tab' });
-        await nextRender();
-
-        expect(overlay.opened).to.be.false;
-      });
-
-      it('should focus next trail item when closed on Tab', async () => {
-        await sendKeys({ press: 'Tab' });
-        await nextRender();
-
-        const nextLink = items[2].shadowRoot.querySelector('[part="link"]');
-        expect(getDeepActiveElement()).to.equal(nextLink);
-      });
-
-      it('should focus previous trail item when closed on Shift+Tab', async () => {
-        await sendKeys({ press: 'Shift+Tab' });
-        await nextRender();
-
-        const prevLink = items[0].shadowRoot.querySelector('[part="link"]');
-        expect(getDeepActiveElement()).to.equal(prevLink);
-      });
-
-      it('should return focus to the overflow button when closed via Escape', async () => {
-        await sendKeys({ press: 'Escape' });
-        await nextRender();
-
-        expect(breadcrumbs.shadowRoot.activeElement).to.equal(button);
-      });
-
-      it('should close the overlay on outside click', async () => {
-        document.body.click();
-        await nextRender();
-
-        expect(overlay.opened).to.be.false;
-      });
     });
   });
 });
